Add files via upload
This is just one of the package files I've created when going through Microsoft's Visual Studio 2019 tutorial on how to create an ETL package.
 https://docs.microsoft.com/en-us/sql/integration-services/ssis-how-to-create-an-etl-package?view=sql-server-ver15

Note, the tutorial isn't perfect and I've already started over 3 times. But that's the beauty of this, right? If it was easy, everyone would be doing it.